# apps/api/src/openapi.yaml
openapi: 3.1.0
info:
  title: ashim API
  version: 0.9.0
  description: |
    REST API for ashim, a self-hosted image processing platform with 30+ tools.

    ## Authentication

    Most endpoints require authentication via one of two methods:

    1. **Session cookie** — Call `POST /api/auth/login` with username and password. The response includes a `token` field. Pass it as `Authorization: Bearer <token>` on subsequent requests.

    2. **API key** — Generate a key via the Settings UI or `POST /api/v1/api-keys`. Keys are prefixed with `si_`. Pass as `Authorization: Bearer si_...`.

    Endpoints marked with a lock icon require authentication. Admin-only endpoints are noted in their description.
  license:
    name: AGPL-3.0
    url: https://github.com/ashim-hq/ashim/blob/main/LICENSE
servers:
  - url: /
    description: Current instance
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token from login or API key (prefixed with si_)
  schemas:
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Invalid image format
    ToolResponse:
      type: object
      properties:
        jobId:
          type: string
          description: Unique job identifier
        downloadUrl:
          type: string
          description: URL to download the processed image
          example: /api/v1/download/abc123/output.png
        originalSize:
          type: integer
          description: Original file size in bytes
        processedSize:
          type: integer
          description: Processed file size in bytes
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/tools/image-to-pdf:
    post:
      tags: [Tools]
      summary: Image to PDF
      description: Convert one or more images into a PDF document.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, settings]
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: One or more image files to include in the PDF
                settings:
                  type: string
                  description: |
                    JSON string with options:
                    - `pageSize` (string, default "A4") — One of: A4, Letter, A3, A5
                    - `orientation` (string, default "portrait") — One of: portrait, landscape
                    - `margin` (number 0-100, default 20) — Page margin in points
      responses:
        "200":
          description: Generated PDF (downloadUrl points to .pdf file)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
...
  /api/v1/tools/favicon:
    post:
      tags: [Tools]
      summary: Favicon generator
      description: Generate a complete favicon set (16px to 512px PNGs, ICO, and manifest.json) from an image. Returns a ZIP file.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file to use as favicon source
      responses:
        "200":
          description: ZIP archive with results
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
...
  /api/v1/tools/find-duplicates:
    post:
      tags: [Tools]
      summary: Find duplicates
      description: Find duplicate images in a set using perceptual hashing.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Multiple image files to check for duplicates
      responses:
        "200":
          description: Duplicate groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items:
                      type: object
                      properties:
                        files:
                          type: array
                          items:
                            type: string
                        similarity:
                          type: number
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
...
  /api/v1/download/{jobId}/{filename}:
    get:
      tags: [Files]
      summary: Download a processed image
      security: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Image binary
          content:
            image/*:
              schema:
                type: string
                format: binary
        "404":
          description: File not found
...